requirements.txt
loguru
pillow
requests